requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0 